gcc 4.3 deprecated features fix: use tr1 for hash_map, const TCHAR* must be used for strings.
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2941 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/test/testall.cpp
@@ -14,6 +14,7 @@
 #endif
 
 #include "test.h"
+#include <stdlib.h>
 
 #include <fcntl.h>
 #ifdef _CL_HAVE_DIRECT_H
@@ -193,7 +194,7 @@ int main(int argc, char *argv[])
     _tprintf(_T("%s\n"), output->buffer);
 
 	if ( times )
-		printf("Tests run in %dms\n\n",CL_NS(util)::Misc::currentTimeMillis()-startTime);
+		printf("Tests run in %dms\n\n", (int32_t)(CL_NS(util)::Misc::currentTimeMillis()-startTime));
 
 exit_point:
 	if ( alltests != NULL )
